moved mitmproxy in a separate view
